Add chart and methods
